refactor: Remove unnecessary sub-folders inside of the components folder
